Sequential CCR

CCRls runs regressions with potentially more covariates than observations. See c_chmod() for the list of models supported.

CCRls(Y, X, kap = 0.1, modclass = "lm", tol = 1e-06, reltol = TRUE,
  rndcov = NULL, report = NULL, ...)

Arguments

Y
vector of dependent variable Y

X
design matrix (without intercept)

kap
maximum number of parameters to estimate in each active sequential step, as a fraction of the less of total number of observations n or number of covariates p. i.e. \(min(n,p)\)

modclass
a string denoting the desired the class of model. See c_chmod for details.

tol
level of tolerance for convergence; default tol=1e-6

reltol
a logical for relative tolerance instead of level. Defaults at TRUE

rndcov
seed for randomising assignment of covariates to partitions; default NULL

report
number of iterations after which to report progress; default NULL

...
additional arguments to be passed to the model

Value

betas parameter estimates (intercept first),

iter number of iterations,

dev increment in the objective function value at convergence

fval objective function value at convergence

Examples

set.seed(14) #Generate data
N = 1000; (bets = rep(-2:2,4)); p = length(bets); X = matrix(rnorm(N*p),N,p)

#>  [1] -2 -1  0  1  2 -2 -1  0  1  2 -2 -1  0  1  2 -2 -1  0  1  2

Y = cbind(1,X)%*%matrix(c(0.5,bets),ncol = 1)
CCRls(Y,X,kap=0.1,modclass="lm",tol=1e-6,reltol=TRUE,rndcov=NULL,report=8)

#> Iter = 8 fval = 171.6381 
#> Iter = 16 fval = -2021.848 
#> Iter = 24 fval = -4238.609 
#> Iter = 32 fval = -6142.191 
#> Iter = 40 fval = -7774.868 
#> Iter = 48 fval = -9587.035 
#> Iter = 56 fval = -11390.67 
#> Iter = 64 fval = -13005.29 
#> Iter = 72 fval = -14775.72 
#> Iter = 80 fval = -16533.27 
#> Iter = 88 fval = -18396.9 
#> Iter = 96 fval = -20283.21 
#> Iter = 104 fval = -21899.35 
#> Iter = 112 fval = -23697.19 
#> Iter = 120 fval = -25467.95 
#> Iter = 128 fval = -27271.29 
#> Iter = 136 fval = -29144.27 

#> $betas
#>   (Intercept)                                                         
#>  5.000000e-01 -2.000000e+00 -1.000000e+00  9.545726e-15  1.000000e+00 
#>                                                                       
#>  2.000000e+00 -2.000000e+00 -1.000000e+00 -2.844662e-15  1.000000e+00 
#>                                                                       
#>  2.000000e+00 -2.000000e+00 -1.000000e+00 -5.262962e-16  1.000000e+00 
#>                                                                       
#>  2.000000e+00 -2.000000e+00 -1.000000e+00 -1.949927e-15  1.000000e+00 
#>               
#>  2.000000e+00 
#> 
#> $iter
#> [1] 139
#> 
#> $dev
#> 'log Lik.' -0.00236367 (df=5)
#> 
#> $fval
#> 'log Lik.' -29814.17 (df=5)
#>